+++ lib/auth.js
@@ -1,4 +1,4 @@
-module.exports = function(app, config) {
+module.exports = function(app, config, redisClient) {
   var passport = require('passport')
     , AtlassianCrowdStrategy = require('passport-atlassian-crowd').Strategy
     , LocalStrategy = require('passport-local').Strategy
@@ -45,12 +45,12 @@ module.exports = function(app, config) {
     authenticationStrategy = 'atlassian-crowd';
     passport.serializeUser(function(user, done) {
       var userId = RegExp('[^/]*$').exec(user.id)||[,null][1];
-      app.locals.redisClient.set("user:"+userId, JSON.stringify(user));
+      redisClient.set("user:"+userId, JSON.stringify(user));
       done(null, userId);
     });
 
     passport.deserializeUser(function(id, done) {
-      app.locals.redisClient.get("user:"+id, function(err, data) {
+      redisClient.get("user:"+id, function(err, data) {
         try {
           var parsedJSON = JSON.parse(data);
         }
