+++ package.json
@@ -21,20 +21,17 @@
   },
   "scripts": {
     "start": "forever start -a -l c/var/log/cloudykangaroo/forever.log -o /var/log/cloudykangaroo/out.log -e /var/log/cloudykangaroo/err.log app.js",
-    "postdeploy": "mkdir -p ./logs",
-    "test": "mocha tests"
+    "test": "mocha --recursive -u tdd --reporter spec"
   },
   "dependencies": {
     "express": "3.4.x",
     "redis": "0.10.x",
     "fakeredis": "0.1.3",
-    "nock": "0.27.x",
     "mocha": "1.17.x",
     "assert": "1.1.x",
     "supertest": "0.9.x",
     "connect-redis": "1.4.x",
     "async": "0.2.x",
-    "colors": "0.6.x",
     "util": "0.10.x",
     "uuid": "1.4.x",
     "express-useragent": "0.0.11",
