+++ test/app.test.js
@@ -3,6 +3,11 @@ process.env.LOG_DIR = './';
 var app = require('../app');
 var request = require('supertest');
 var assert = require('assert');
+var port = 3333;
+
+// Test Writing reference:
+// https://github.com/visionmedia/supertest
+// https://github.com/visionmedia/express/tree/master/test
 
 describe('GET /', function(){
   it('should return code 302', function(done){
@@ -16,26 +21,12 @@ describe('GET /api/v1/ubersmith/devices/deviceid/10020', function(){
   it('should return code 200', function(done){
     request(app)
       .get('/api/v1/ubersmith/devices/deviceid/10020')
-      .expect(200)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect(200, done);
   });
   it('should respond with json', function(done){
     request(app)
       .get('/api/v1/ubersmith/devices/deviceid/10020')
-      .expect('Content-Type', /json/)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect('Content-Type', /json/, done);
   });
 });
 
@@ -43,26 +34,12 @@ describe('GET /api/v1/ubersmith/devices/deviceid/10020/tickets', function(){
   it('should return code 200', function(done){
     request(app)
       .get('/api/v1/ubersmith/devices/deviceid/10020/tickets')
-      .expect(200)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect(200, done);
   });
   it('should respond with json', function(done){
     request(app)
       .get('/api/v1/ubersmith/devices/deviceid/10020/tickets')
-      .expect('Content-Type', /json/)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect('Content-Type', /json/, done);
   });
 });
 
@@ -70,65 +47,30 @@ describe('GET /api/v1/ubersmith/devices/hostname', function(){
   it('should return code 200', function(done){
     request(app)
       .get('/api/v1/ubersmith/devices/hostname')
-      .expect(200)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
-    it('should respond with json', function(done){
-      request(app)
-        .get('/api/v1/ubersmith/devices/hostname')
-        .expect('Content-Type', /json/)
-        .end(function(err, res){
-          if(err) {
-            done(err);
-          } else {
-            done();
-          }
-        });
-    });
-  })
+      .expect(200, done);
+  });
+  it('should respond with json', function(done){
+    request(app)
+      .get('/api/v1/ubersmith/devices/hostname')
+      .expect('Content-Type', /json/, done);
+  });
 });
 
 describe('GET /api/v1/ubersmith/devices/hostname/jsklskwtrs-engage05.unittest.us', function(){
   it('should return code 200', function(done) {
     request(app)
       .get('/api/v1/ubersmith/devices/hostname/jsklskwtrs-engage05.unittest.us')
-      .expect(200)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect(200, done);
   });
   it('should respond with json', function(done){
     request(app)
       .get('/api/v1/ubersmith/devices/hostname/jsklskwtrs-engage05.unittest.us')
-      .expect('Content-Type', /json/)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect('Content-Type', /json/, done);
   });
-  it('should return the same sent params concatenated', function(done) {
+  it('should return jsklskwtrs-engage05', function(done) {
     request(app)
       .get('/api/v1/ubersmith/devices/hostname/jsklskwtrs-engage05.unittest.us')
-      .expect(200, {"aaData":[{"domain":"jaskolski-waters.com","deviceID":10023,"typeGroupID":"1","shortname":"jsklskwtrs","ipAddr":"10.50.33.47","typeGroupName":"Servers","clientID":1003,"name":"jsklskwtrs-engage05","company":"Jaskolski-Waters","management_level":"managed","device_status":"Active","type":"Servers","dev_desc":"jsklskwtrs-engage05","dev":10023,"client_id":1003,"clientid":1003,"label":"jsklskwtrs-servers","active":1,"location":"SJC"}]})
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect(200, {"aaData":[{"domain":"jaskolski-waters.com","deviceID":10023,"typeGroupID":"1","shortname":"jsklskwtrs","ipAddr":"10.50.33.47","typeGroupName":"Servers","clientID":1003,"name":"jsklskwtrs-engage05","company":"Jaskolski-Waters","management_level":"managed","device_status":"Active","type":"Servers","dev_desc":"jsklskwtrs-engage05","dev":10023,"client_id":1003,"clientid":1003,"label":"jsklskwtrs-servers","active":1,"location":"SJC"}]}, done)
   });
 });
 
@@ -136,26 +78,12 @@ describe('GET /api/v1/ubersmith/clients/clientid/1022', function(){
   it('should return code 200', function(done){
     request(app)
       .get('/api/v1/ubersmith/clients/clientid/1022')
-      .expect(200)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect(200, done);
   });
   it('should respond with json', function(done){
     request(app)
       .get('/api/v1/ubersmith/clients/clientid/1022')
-      .expect('Content-Type', /json/)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect('Content-Type', /json/, done);
   });
 })
 
@@ -163,26 +91,12 @@ describe('GET /api/v1/ubersmith/devices/devgroupid/1', function(){
   it('should return code 200', function(done){
     request(app)
       .get('/api/v1/ubersmith/devices/devgroupid/1')
-      .expect(200)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect(200, done);
   });
   it('should respond with json', function(done){
     request(app)
       .get('/api/v1/ubersmith/devices/devgroupid/1')
-      .expect('Content-Type', /json/)
-      .end(function(err, res){
-        if(err) {
-          done(err);
-        } else {
-          done();
-        }
-      });
+      .expect('Content-Type', /json/, done);
   });
 });
